bot/requirements.txt
python-dotenv
aiogram
psycopg2-binary
requests
googletrans==4.0.2